pyproject.toml
[project]
name = "dj1"
version = "0.1.0"
description = "some real niqqa shit"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    # ── Audio Analysis ─────────────────────────────
    "librosa>=0.10.2",
    "aubio>=0.4.9",
    "soundfile>=0.12.1",
    "numpy>=1.26.4",

    # ── Audio Manipulation & Effects ───────────────
    "pydub>=0.25.1",
    "pedalboard>=0.9.9",
    "pyrubberband>=0.3.0",

    # ── Metadata ───────────────────────────────────
    "mutagen>=1.47.0",

    # ── Spotify Integration ────────────────────────
    "spotipy>=2.24.0",
    "python-dotenv>=1.0.1",

    # ── Matching & Utilities ───────────────────────
    "thefuzz>=0.22.1",
    "python-levenshtein>=0.25.1",
    "joblib>=1.4.2",
    "pyyaml>=6.0.2",
    "tqdm>=4.66.4",
    "colorama>=0.4.6",
    "rich>=13.7.1",
]

[project.scripts]
autodj = "main:run_pipeline"

[dependency-groups]
    dev = [
     "pytest>=8.0.0",
     "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modules", "utils"]

[tool.uv]
package = true